move aggregate related things to aggr crate
