+++ datafusion/physical-plan/Cargo.toml
@@ -46,6 +46,7 @@ datafusion-common = { workspace = true, default-features = true }
 datafusion-common-runtime = { workspace = true, default-features = true }
 datafusion-execution = { workspace = true }
 datafusion-expr = { workspace = true }
+datafusion-functions-aggregate = { workspace = true }
 datafusion-physical-expr = { workspace = true, default-features = true }
 futures = { workspace = true }
 half = { workspace = true }
